+++ package.json
@@ -31,7 +31,7 @@
     "react-calendar": "^5.0.0",
     "react-datepicker": "^7.3.0",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.52.1",
+    "react-hook-form": "^7.52.2",
     "react-icons": "^5.2.1",
     "react-kakao-maps-sdk": "^1.1.27",
     "react-router-dom": "^6.25.1",
