+++ vite.config.js
@@ -23,7 +23,8 @@ export default defineConfig({
       { find: '@zustands', replacement: resolve(__dirname, 'src/zustands') },
       { find: '@axios', replacement: resolve(__dirname, 'src/axios') },
       { find: '@assets', replacement: resolve(__dirname, 'src/assets') },
-      { find: '@Header', replacement: resolve(__dirname, 'src/components/Header') }, // 수정된 부분
+      { find: '@Header', replacement: resolve(__dirname, 'src/components/Header') },
+      { find: '@auth', replacement: resolve(__dirname, 'src/components/auth') },
       { find: '@', replacement: resolve(__dirname, 'src') },
     ],
   },
